% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPPSims_CreateResultsTibbles.R
\name{clusterPPSims_CreateResultsTibbles}
\alias{clusterPPSims_CreateResultsTibbles}
\title{Create parent and cluster tibbles from cluster point process simulation object}
\usage{
clusterPPSims_CreateResultsTibbles(res)
}
\value{
list with elements "parents" and "clusters", each of which is a tbl_df object
with the first containing cluster parent locations by simulation and the second
containing the locations of the points in the clusters, by simulation.
}
\description{
Create parent and cluster tibbles from cluster point process simulation object
}
