% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSims_GenerateNormalCluster.R
\name{clusterSims_GenerateNormalCluster}
\alias{clusterSims_GenerateNormalCluster}
\title{Generate anisotropic cluster of points based on a normal 2d distribution}
\usage{
clusterSims_GenerateNormalCluster(
  x0,
  y0,
  n,
  sigx,
  sigy,
  theta,
  sign = 0,
  sigt = 0
)
}
\arguments{
\item{x0}{\itemize{
\item x-coordinate of cluster "center" in "map" coordinates
}}

\item{y0}{\itemize{
\item y-coordinate of cluster "center" in "map" coordinates
}}

\item{n}{\itemize{
\item number of points to generate at randomly-selected locations
}}

\item{sigx}{\itemize{
\item std. deviation of displacements along main axis
}}

\item{sigy}{\itemize{
\item std. deviation of displacements along secondary (perpindicular) axis
}}

\item{theta}{\itemize{
\item rotation of main axis relative to the "map"'s coordinate system x-axis
}}

\item{sign}{\itemize{
\item std. deviation of expected number in cluster
}}

\item{sigt}{\itemize{
\item std. deviation of cluster axis rotations
}}
}
\value{
list with elements "x" and "y", vectors giving coordinates of
generated points in the "map" coordinate system.
}
\description{
Function used for anisotropic cluster generation based on a normal 2d distribution.
}
\details{
For each cluster"n" points are generated at locations relative to a cluster "center"
using a bivariate normal distribution with 2d axes aligned along the principal
direction of variation and perpindicular to it. The displacement variances along
these axes are given by $sigx^2$ and $sigy^2$. The generated displacements are then
rotated by "theta" and added to the cluster "center" coordinates to yield locations
in "map" (real world) coordinates.
}
