% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isa_ppplist.R
\name{isa_ppplist}
\alias{isa_ppplist}
\title{Check if object is a ppplist object}
\usage{
isa_ppplist(x)
}
\arguments{
\item{x}{\itemize{
\item object to check
}}
}
\value{
TRUE or FALSE
}
\description{
Funciton to check if object is a ppplist object.
}
