% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstSamplingInfo.R
\docType{data}
\name{lstSamplingInfo}
\alias{lstSamplingInfo}
\title{lstSamplingInfo}
\format{
A list with elements:
\describe{
\item{\code{tdBSFRF}}{mean total distance towed for BSFRF tows}
\item{\code{nwBSFRF}}{mean net width for BSFRF tows}
\item{\code{asBSFRF}}{mean area swept for BSFRF tows}
\item{\code{tdNMFS}}{mean total distance towed for NMFS tows}
\item{\code{nwNMFS}}{mean net width for NMFS tows}
\item{\code{asNMFS}}{mean area swept for NMFS tows}
}
}
\usage{
lstSamplingInfo
}
\description{
mean sampling information for
BSFRF and NMFS vessels during SBS tow operations.
}
\keyword{datasets}
