% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSBSResults_Hists1.R
\name{plotSBSResults_Hists1}
\alias{plotSBSResults_Hists1}
\title{Create histogram plots of side-by-side (SBS) haul study results}
\usage{
plotSBSResults_Hists1(dfr, type_)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with results from ??
}}

\item{type_}{\itemize{
\item flag indicating which type of results to plot
}}
}
\value{
list with two plots
}
\description{
Function to create histogram plots of side-by-side (SBS) haul study results.
}
\details{
none.
}
