% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppSims_CalcStats.R
\name{ppSims_CalcStats}
\alias{ppSims_CalcStats}
\title{Compute stats for simulated point process}
\usage{
ppSims_CalcStats(resp)
}
\value{
a dataframe
}
\description{
Compute stats for simulated point process
}
