% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppSims_CreateResultsTibbles.R
\name{ppSims_CreateResultsTibbles}
\alias{ppSims_CreateResultsTibbles}
\title{Create result tibbles from a list of point process simulations}
\usage{
ppSims_CreateResultsTibbles(res)
}
\arguments{
\item{res}{\itemize{
\item ppplist object
}}
}
\value{
pp_sims_tbls object (a list with elements "parents" and "clusters")
}
\description{
Function to create result tibbles from a list of point process simulations.
}
\details{
Delegates to \code{\link[=simplePPSims_CreateResultsTibbles]{simplePPSims_CreateResultsTibbles()}} or \code{\link[=clusterPPSims_CreateResultsTibbles]{clusterPPSims_CreateResultsTibbles()}}
depending on the attributes of the first simulation in the input ppplist.
}
