% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppSims_MakePlots.R
\name{ppSims_MakePlots}
\alias{ppSims_MakePlots}
\title{Make plots of the results of a point process simulation}
\usage{
ppSims_MakePlots(
  resp,
  all_parents = FALSE,
  contour_type = "none",
  bins = NULL,
  binwidth = NULL,
  breaks = NULL
)
}
\arguments{
\item{resp}{\itemize{
\item pp_sims_tbls or ppplist object
}}

\item{all_parents}{\itemize{
\item T/F to plot location of all parents (including ones with no children)
}}

\item{contour_type}{\itemize{
\item string indicating contour type ("none", "density", "ndensity", or "count")
}}

\item{bins}{\itemize{
\item number of contour bins (passed to geom_contour)
}}

\item{binwidth}{\itemize{
\item width of contour bins (passed to geom_contour)
}}

\item{breaks}{\itemize{
\item vector of contour bin cutpoints (passed to geom_contour, overrides bins & binwidth)
}}
}
\value{
list with ggplot2 objects "clusters", "parents",
and (if "all_parents" is TRUE) "all_parents".
}
\description{
Make plots of the results of a point process simulation
}
