% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratcount_AsTbl.R
\name{quadratcount_AsTbl}
\alias{quadratcount_AsTbl}
\title{Create a dataframe from a quadratcount object}
\usage{
quadratcount_AsTbl(qc)
}
\arguments{
\item{qc}{\itemize{
\item the quandratcount object
}}
}
\value{
a dataframe
}
\description{
Function to create a dataframe from a quadratcount object.
}
\details{
\itemize{
\item none
}
}
