% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppSims_CalcStats.R
\name{ppSims_CalcStats}
\alias{ppSims_CalcStats}
\title{Compute stats for simulated point process}
\usage{
ppSims_CalcStats(resp)
}
\arguments{
\item{resp}{\itemize{
\item a ppplist or pp_sims_tbls object
}}
}
\value{
a dataframe
}
\description{
Function to compute statistics for a simulated point process.
}
